refactor: move transcript pop into separate function
